requirements.txt
# Knowledge Extraction Pipeline (KEP) Dependencies
# Core KEP requirements
pyyaml>=6.0
pathlib
requests>=2.28.0
python-dotenv>=0.19.0

# LLM Providers
ibm-watsonx-ai>=0.2.0
ibm-cloud-sdk-core>=3.16.0

# PDF Processing & Document Conversion
#docling
#docling-core

# NLP Processing
nltk>=3.8

#Tika
tika

#PyMuPDF for PDF text extraction
pytesseract 
pymupdf

# Image Processing (for SmolDocling)
pillow>=9.0.0
#mlx-vlm>=0.0.8

# Jupyter Lab Environment (Optional - only if using notebooks)
# jupyterlab>=4.0.0
# notebook>=6.5.0
# ipywidgets>=8.0.0

# Data Processing
pandas>=1.5.0
numpy>=1.24.0

# Utilities
tqdm>=4.64.0
colorama>=0.4.6
rich>=13.0.0

# Development & Testing
pytest>=7.0.0
black>=22.0.0
flake8>=5.0.0
mypy>=1.0.0

# Additional utilities for text processing
regex>=2022.10.31
beautifulsoup4>=4.11.0
markdownify>=0.11.0

# Graph visualization
networkx>=3.0
pyvis>=0.3.2